# backend/service-core/src/main/resources/oas/apis.yaml
openapi: 3.0.3
info:
  title: Budget App API
  description: APIs for managing the budget control app.
  version: 1.0.0
paths:
...
  /organizations/{name}/resources:
    get:
      summary: Get all resources
      operationId: getAllResourcesByOrganizationName
      tags:
        - Resources
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Resource'
    post:
      summary: Create a new resource
      operationId: createResourceByOrganizationName
      tags:
        - Resources
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resource'
      responses:
        '201':
          description: Resource created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
...
  /organizations/{name}/clients/{clientId}/projects/{projectId}/resources:
    get:
      summary: Get all resources associated to a client and a project
      operationId: getAllResourceByClientIdAndProjectId
      tags:
        - Resources
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Resource'
    post:
      summary: Associate a resource to a project
      operationId: associateResourceToProjectByOrganizationNameAndClientId
      tags:
        - Resources
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resource'
      responses:
        '200':
          description: Resource created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  resource:
                    $ref: '#/components/schemas/Resource'
                  project:
                    $ref: '#/components/schemas/Project'
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
          description: The id of the project
        name:
          type: string
          description: The name of the project
        revenues:
          type: number
          format: double
          description: The revenue expected for the project
        cci:
          type: number
          format: double
          description: The cci expected for the project
        startDate:
          type: string
          example: "2021-01-01"
        endDate:
          type: string
          example: "2021-01-01"
...
    Resource:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        level:
          enum:
            - Analyst
            - Consultant
            - Manager
            - Senior Manager
            - Managing Director
        lcr:
          type: number
          format: double
        chargeability:
          type: number
          format: double